Restructure as python package
