+++ metersink/__main__.py
@@ -1,0 +1,2 @@
+from metersink.api import main
+main()
